# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tài liệu API E-commerce Fashion",
        "description": "Hợp đồng API giữa Vue.js Frontend và Laravel Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/customer-addresses": {
            "get": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Lấy danh sách địa chỉ khách hàng",
                "operationId": "54ec4e543fd8d7157854ef51e8ade3e2",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Lọc theo ID khách hàng",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm kiếm theo tên người nhận hoặc số điện thoại",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Thêm địa chỉ mới",
                "operationId": "5b08800375082c3893d949d2eaf62cd9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "receiver_name",
                                    "receiver_phone",
                                    "detail_address"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "receiver_name": {
                                        "type": "string"
                                    },
                                    "receiver_phone": {
                                        "type": "string"
                                    },
                                    "province": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "district": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ward": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "detail_address": {
                                        "type": "string"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "422": {
                        "description": "Lỗi validate dữ liệu"
                    }
                }
            }
        },
        "/api/admin/customer-addresses/{customer_address}": {
            "get": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Xem chi tiết địa chỉ",
                "operationId": "5d4d9e595cd3633e9cdfb18500504630",
                "parameters": [
                    {
                        "name": "customer_address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Cập nhật địa chỉ",
                "operationId": "d227203dd3f553b6de032a6907198d2f",
                "parameters": [
                    {
                        "name": "customer_address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "receiver_name",
                                    "receiver_phone",
                                    "detail_address"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "receiver_name": {
                                        "type": "string"
                                    },
                                    "receiver_phone": {
                                        "type": "string"
                                    },
                                    "province": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "district": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ward": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "detail_address": {
                                        "type": "string"
                                    },
                                    "is_default": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "422": {
                        "description": "Lỗi validate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CustomerAddresses"
                ],
                "summary": "Xóa địa chỉ",
                "operationId": "2f8ebb1cf529c090345622925f6bd376",
                "parameters": [
                    {
                        "name": "customer_address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
...
}